Refactor socket handling into modules (similar to file). Simplify getaddrinfo_detour.
