+++ mirrord-layer/src/lib.rs
@@ -15,7 +15,6 @@ use common::{
 };
 use ctor::ctor;
 use envconfig::Envconfig;
-use error::LayerError;
 use file::OPEN_FILES;
 use frida_gum::{interceptor::Interceptor, Gum};
 use futures::{SinkExt, StreamExt};
@@ -27,7 +26,7 @@ use mirrord_protocol::{
     OpenFileRequest, OpenFileResponse, OpenRelativeFileRequest, ReadFileRequest, ReadFileResponse,
     SeekFileRequest, SeekFileResponse, WriteFileRequest, WriteFileResponse,
 };
-use sockets::SOCKETS;
+use socket::SOCKETS;
 use tcp::TcpHandler;
 use tcp_mirror::TcpMirrorHandler;
 use tokio::{
@@ -48,7 +47,7 @@ mod error;
 mod file;
 mod macros;
 mod pod_api;
-mod sockets;
+mod socket;
 mod tcp;
 mod tcp_mirror;
 
@@ -266,7 +265,6 @@ async fn handle_hook_message(
         }
     }
 }
-
 
 #[allow(clippy::too_many_arguments)]
 async fn handle_daemon_message(
@@ -507,7 +505,7 @@ fn enable_hooks(enabled_file_ops: bool) {
 
     hook!(interceptor, "close", close_detour);
 
-    sockets::enable_socket_hooks(&mut interceptor);
+    socket::hooks::enable_socket_hooks(&mut interceptor);
 
     if enabled_file_ops {
         file::hooks::enable_file_hooks(&mut interceptor);
